Rename main to home
