+++ app/src/main/java/ch/admin/bag/dp3t/MainActivity.java
@@ -21,7 +21,7 @@ import androidx.lifecycle.ViewModelProvider;
 
 import org.dpppt.android.sdk.DP3T;
 
-import ch.admin.bag.dp3t.main.HomeFragment;
+import ch.admin.bag.dp3t.home.HomeFragment;
 import ch.admin.bag.dp3t.networking.ConfigWorker;
 import ch.admin.bag.dp3t.onboarding.OnboardingActivity;
 import ch.admin.bag.dp3t.reports.ReportsFragment;
