+++ app/src/main/java/ch/admin/bag/dp3t/home/model/NotificationState.java
